🎨 From now on the XES_Importer handles extended traces.
